+++ SkyEdge - Springboot/src/main/java/SkyEdge/repository/UserRepository.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import SkyEdge.model.MyUserDetails;
+import SkyEdge.model.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<MyUserDetails, Integer> {
-    Optional<MyUserDetails> findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findByUsername(String username);
 }
