+++ SkyEdge - Springboot/src/main/java/SkyEdge/service/AuthenticationService.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import SkyEdge.model.MyUserDetails;
 import SkyEdge.model.Role;
+import SkyEdge.model.User;
 import SkyEdge.repository.RoleRepository;
 import SkyEdge.repository.UserRepository;
 import jakarta.transaction.Transactional;
@@ -25,14 +25,14 @@ public class AuthenticationService {
     @Autowired
     private PasswordEncoder passwordEncoder;
 
-    public MyUserDetails registerUser(String username, String password) {
+    public User registerUser(String username, String password) {
         String encodedPassword = passwordEncoder.encode(password);
         Role userRole = roleRepository.findByAuthority("ADMIN").get();
 
         Set<Role> authorities = new HashSet<>();
         authorities.add(userRole);
         System.out.println("role added");
-        return userRepository.save(new MyUserDetails(username, encodedPassword, authorities));
+        return userRepository.save(new User(username, encodedPassword, authorities));
     }
 
 }
