+++ SkyEdge - Springboot/src/main/resources/templates/register.html
@@ -6,12 +6,12 @@
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
 	<title>Register</title>
 
-	<link rel="shortcut icon" href="../static/assets/favicon.png" type="image/x-icon">
+	<link rel="shortcut icon" th:href="@{/assets/favicon.png}" type="image/x-icon">
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css">
-	<link rel="stylesheet" href="../static/pages/Register/register.css">
-	<link rel="stylesheet" href="../static/assets/css/colors.css">
-	<link rel="stylesheet" href="../static/assets/css/typo.css">
-	<link rel="stylesheet" href="../static/displays/AccountDisplay/account.css">
+	<link rel="stylesheet" th:href="@{/pages/Register/register.css}">
+	<link rel="stylesheet" th:href="@{/assets/css/colors.css}">
+	<link rel="stylesheet" th:href="@{/assets/css/typo.css}">
+	<link rel="stylesheet" th:href="@{/displays/AccountDisplay/account.css}">
 </head>
 
 <body>
@@ -52,15 +52,15 @@
 						<button type="submit" class="btn btn-primary custom-btn-width btn-signup">Sign up</button>
 					</div>
 
-					<p class="mt-3">Have an account? <a class="log-in-btn" href="./login.html">Log in</a></p>
+					<p class="mt-3">Have an account? <a class="log-in-btn" href="/login">Log in</a></p>
 				</form>
 			</div>
 		</div>
 		<div class="column">
-			<img src="../static/assets/Login_background.png" class="image" alt="Background">
+			<img th:src="@{/assets/Login_background.png}" class="image" alt="Background">
 		</div>
 	</div>
-	<script src="./pages/Register/register.js"></script>
+	<script th:src="@{/pages/Register/register.js}"></script>
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
 </body>
